+++ framework/Response/ResponseInterface.php
@@ -4,8 +4,11 @@ namespace Framework\Response;
 
 interface ResponseInterface
 {
-    public function setStatusCode(String $status_code) : void;
-    public function setHeaders(Array $headers) : void;
-    public function setContent($headers) : void;
-    public function send() : void;
+    public function getStatusCode() : String;
+    public function setStatusCode(String $status_code) : Void;
+    public function getHeaders() : Array;
+    public function setHeaders(Array $headers) : Void;
+    public function getContent();
+    public function setContent($headers) : Void;
+    public function send() : Void;
 }