+++ framework/Router/RouteInterface.php
@@ -6,11 +6,13 @@ use Framework\Request\RequestInterface;
 
 interface RouteInterface
 {
-    public function setUri(String $uri) : RouteInterface;
-    public function getUri() : String;
+    public function setPattern(String $uri) : RouteInterface;
+    public function getPattern() : String;
     public function setMethod(String $method) : RouteInterface;
     public function getMethod() : String;
     public function setController(String $controller) : RouteInterface;
     public function getController() : String;
+    public function getParameters() : Array;
+    public function where(Array $conditions) : RouteInterface;
     public function matchesRequest(RequestInterface $request) : Bool;
 }